+++ apps/docs/src/app/core/component-docs/date-picker/date-picker-docs.module.ts
@@ -25,6 +25,7 @@ import { FdDatetimeModule } from '@fundamental-ngx/core/datetime';
 import { DatePickerModule } from '@fundamental-ngx/core/date-picker';
 import { PipeModule } from '@fundamental-ngx/core/utils';
 import { DatePickerUpdateOnBlurExampleComponent } from './examples/date-picker-update-on-blur-example.component';
+import { DatePickerDisableFocusScrollExampleComponent } from './examples/date-picker-disable-focus-scroll-example.component';
 
 const routes: Routes = [
     {
@@ -64,7 +65,8 @@ const routes: Routes = [
         DatePickerComplexI18nExampleComponent,
         DatePickerDisableFuncExampleComponent,
         DatePickerRangeDisabledExampleComponent,
-        DatePickerUpdateOnBlurExampleComponent
+        DatePickerUpdateOnBlurExampleComponent,
+        DatePickerDisableFocusScrollExampleComponent
     ]
 })
 export class DatePickerDocsModule {}