+++ apps/docs/src/app/core/documentation/core-documentation-data.ts
@@ -111,7 +111,8 @@ export const utilities: SectionInterfaceContent[] = [
     { url: 'core/infinite-scroll', name: 'Infinite Scroll' },
     { url: 'core/initial-focus', name: 'Initial Focus' },
     { url: 'core/scroll-spy', name: 'Scroll Spy' },
-    { url: 'core/theme-switcher', name: 'Theme Switcher API' }
+    { url: 'core/theme-switcher', name: 'Theme Switcher API' },
+    { url: 'core/theming', name: 'Theming' }
 ];
 
 export const adapters: SectionInterfaceContent[] = [