+++ apps/docs/src/app/documentation/core-helpers/toolbar/toolbar.component.ts
@@ -1,21 +1,27 @@
-import { Component, EventEmitter, OnDestroy, OnInit, Output, ViewChild } from '@angular/core';
+import { Component, EventEmitter, Inject, OnDestroy, OnInit, Output, ViewChild } from '@angular/core';
+import { CompleteThemeDefinition, ThemingService } from '@fundamental-ngx/core/theming';
 import { environment } from '../../../../environments/environment';
 import { Router, ActivatedRoute } from '@angular/router';
-import { Libraries } from '../../utilities/libraries';
+import { CURRENT_LIB, Libraries } from '../../utilities/libraries';
 
-import { SafeResourceUrl } from '@angular/platform-browser';
-import { DocsThemeService } from '../../services/docs-theme.service';
-import { fromEvent, Subject } from 'rxjs';
+import { DomSanitizer, SafeResourceUrl } from '@angular/platform-browser';
+import { filter, fromEvent, Subject } from 'rxjs';
 import { debounceTime, startWith, takeUntil } from 'rxjs/operators';
 import { MenuComponent, MenuKeyboardService } from '@fundamental-ngx/core/menu';
-import { ContentDensity, ContentDensityService, ThemesService } from '@fundamental-ngx/core/utils';
+import { ContentDensity, ContentDensityService } from '@fundamental-ngx/core/utils';
 import { ShellbarMenuItem, ShellbarSizes } from '@fundamental-ngx/core/shellbar';
+
+const urlContains = (themeName: string, search: string): boolean => themeName.toLowerCase().includes(search);
+
+const isHcb = (themeName: string): boolean => urlContains(themeName, 'hcb');
+const isHcw = (themeName: string): boolean => urlContains(themeName, 'hcw');
+const isDark = (themeName: string): boolean => urlContains(themeName, 'dark');
 
 @Component({
     selector: 'fd-docs-toolbar',
     templateUrl: './toolbar.component.html',
     styleUrls: ['./toolbar.component.scss'],
-    providers: [MenuKeyboardService, ThemesService]
+    providers: [MenuKeyboardService]
 })
 export class ToolbarDocsComponent implements OnInit, OnDestroy {
     @Output()
@@ -24,12 +30,13 @@ export class ToolbarDocsComponent implements OnInit, OnDestroy {
     @ViewChild('themeMenu')
     themeMenu: MenuComponent;
 
-    cssUrl: SafeResourceUrl;
-    customCssUrl: SafeResourceUrl;
+    highlightJsThemeCss: SafeResourceUrl;
 
     library: Libraries;
 
     size: ShellbarSizes = 'm';
+
+    themes: CompleteThemeDefinition[];
 
     version = {
         id: environment.version,
@@ -59,27 +66,33 @@ export class ToolbarDocsComponent implements OnInit, OnDestroy {
         }
     ];
 
-    themes = this._themesService.themes;
-
     /** An RxJS Subject that will kill the data stream upon destruction (for unsubscribing)  */
     private readonly _onDestroy$: Subject<void> = new Subject<void>();
 
     constructor(
         private _routerService: Router,
-        private _themesService: ThemesService,
-        private _docsThemeService: DocsThemeService,
         private _contentDensityService: ContentDensityService,
-        _route: ActivatedRoute
+        private _themingService: ThemingService,
+        @Inject(CURRENT_LIB) private _currentLib: Libraries,
+        private _route: ActivatedRoute,
+        private _domSanitizer: DomSanitizer
     ) {
-        this.library = _route.snapshot.data.library || 'core';
+        this._themingService.init();
+        this.library = this._route.snapshot.data.library || 'core';
 
-        this._docsThemeService.onThemeChange.pipe(takeUntil(this._onDestroy$)).subscribe((theme) => {
-            this.cssUrl = theme.themeUrl;
-            this.customCssUrl = theme.customThemeUrl;
-        });
+        this._themingService.currentTheme
+            .pipe(
+                takeUntil(this._onDestroy$),
+                filter((theme) => !!theme)
+            )
+            .subscribe((theme) => {
+                this.updateHighlightTheme(theme?.id as string);
+            });
     }
 
     ngOnInit(): void {
+        this.themes = this._themingService.getThemes();
+
         this.versions = [
             { id: '0.33.2', url: 'https://620423a8f2458b000724fd5f--fundamental-ngx.netlify.app/' },
             { id: '0.32.0', url: 'https://6130e294b2dc5c00086828de--fundamental-ngx.netlify.app/' },
@@ -108,10 +121,6 @@ export class ToolbarDocsComponent implements OnInit, OnDestroy {
 
         this.versions.unshift(this.version);
 
-        if (!(this.cssUrl && this.customCssUrl)) {
-            this.selectTheme(this.themes[0].id);
-        }
-
         fromEvent(window, 'resize')
             .pipe(startWith(1), debounceTime(60), takeUntil(this._onDestroy$))
             .subscribe(() => (this.size = this._getShellbarSize()));
@@ -122,18 +131,33 @@ export class ToolbarDocsComponent implements OnInit, OnDestroy {
         this._onDestroy$.complete();
     }
 
-    selectTheme(selectedTheme: string): void {
-        this.cssUrl = this._themesService.setTheme(selectedTheme);
-        this.customCssUrl = this._themesService.setCustomTheme(selectedTheme);
+    updateHighlightTheme(themeName: string): void {
+        let theme = 'googlecode.css';
+        if (isHcb(themeName)) {
+            theme = 'a11y-dark.css';
+        } else if (isHcw(themeName)) {
+            theme = 'a11y-light.css';
+        } else if (isDark(themeName)) {
+            theme = 'tomorrow-night.css';
+        }
+        this.highlightJsThemeCss = this.trustedResourceUrl(`assets/highlight-js-styles/${theme}`);
     }
 
     selectVersion(version: any): void {
         window.open(version.url, '_blank');
     }
 
+    selectTheme(themeId: string): void {
+        this._themingService.setTheme(themeId);
+        this.updateHighlightTheme(themeId);
+    }
+
     selectDensity(density: ContentDensity): void {
         this._contentDensityService.contentDensity.next(density);
     }
+
+    private trustedResourceUrl = (url: string): SafeResourceUrl =>
+        this._domSanitizer.bypassSecurityTrustResourceUrl(url);
 
     private _getShellbarSize(): ShellbarSizes {
         const width = window.innerWidth;