+++ libs/core/src/lib/link/link.module.ts
@@ -2,10 +2,11 @@ import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 
 import { LinkComponent } from './link.component';
+import { PortalModule } from '@angular/cdk/portal';
 
 @NgModule({
     declarations: [LinkComponent],
-    imports: [CommonModule],
+    imports: [CommonModule, PortalModule],
     exports: [LinkComponent]
 })
 export class LinkModule {}