+++ libs/core/src/lib/list/list.tokens.ts
@@ -1,0 +1,3 @@
+import { InjectionToken } from '@angular/core';
+
+export const FD_LIST = new InjectionToken('List component');