+++ libs/core/src/lib/shellbar/shellbar.module.ts
@@ -1,5 +1,6 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
+import { ActionSheetModule } from '@fundamental-ngx/core/action-sheet';
 import { ShellbarComponent } from './shellbar.component';
 import { ProductMenuComponent } from './product-menu/product-menu.component';
 import { PopoverModule } from '@fundamental-ngx/core/popover';
@@ -18,7 +19,7 @@ import { DragAndDropModule } from '@fundamental-ngx/core/utils';
 import { DragDropModule } from '@angular/cdk/drag-drop';
 import { ProductSwitchModule } from '@fundamental-ngx/core/product-switch';
 import { ShellbarSidenavDirective } from './shellbar-sidenav.directive';
-import { ShellbarActionsMobileComponent } from './shellbar-actions/shellbar-actions-mobile.component';
+import { ShellbarActionsMobileComponent } from './shellbar-actions-mobile/shellbar-actions-mobile.component';
 import { AvatarModule } from '@fundamental-ngx/core/avatar';
 
 @NgModule({
@@ -38,6 +39,7 @@ import { AvatarModule } from '@fundamental-ngx/core/avatar';
         CommonModule,
         PopoverModule,
         MenuModule,
+        ActionSheetModule,
         ButtonModule,
         DragAndDropModule,
         DragDropModule,