+++ libs/core/src/lib/theming/tokens.ts
@@ -1,0 +1,4 @@
+import { InjectionToken } from '@angular/core';
+import { ThemingConfig } from './interfaces/theming-config.interface';
+
+export const THEMING_CONFIG_TOKEN = new InjectionToken<Partial<ThemingConfig>>('FdThemingConfig');