+++ libs/core/src/lib/title/title.component.ts
@@ -4,6 +4,10 @@ import { Nullable } from '@fundamental-ngx/core/shared';
 
 export type HeaderSizes = 1 | 2 | 3 | 4 | 5 | 6;
 
+export abstract class TitleToken {
+    abstract elementRef: ElementRef;
+}
+
 @Component({
     // eslint-disable-next-line
     selector: 'h1[fd-title], h2[fd-title], h3[fd-title], h4[fd-title], h5[fd-title], h6[fd-title]',
@@ -15,9 +19,10 @@ export type HeaderSizes = 1 | 2 | 3 | 4 | 5 | 6;
     },
     styleUrls: ['./title.component.scss'],
     changeDetection: ChangeDetectionStrategy.OnPush,
-    encapsulation: ViewEncapsulation.None
+    encapsulation: ViewEncapsulation.None,
+    providers: [{ provide: TitleToken, useExisting: TitleComponent }]
 })
-export class TitleComponent implements OnInit {
+export class TitleComponent extends TitleToken implements OnInit {
     /** The size of the header */
     _headerSize: Nullable<HeaderSizes> = null;
 
@@ -41,11 +46,18 @@ export class TitleComponent implements OnInit {
     private _appliedHeaderSize: number;
 
     /** @hidden */
-    constructor(private _elementRef: ElementRef) {}
+    constructor(private _elementRef: ElementRef) {
+        super();
+    }
 
     /** @hidden */
     ngOnInit(): void {
         this._setHeaderSize();
+    }
+
+    /** returns the reference to the title element */
+    get elementRef(): ElementRef {
+        return this._elementRef;
     }
 
     /** @hidden */