+++ libs/platform/src/lib/icon-tab-bar/components/icon-tab-bar-process-type/icon-tab-bar-process-type.component.ts
@@ -89,7 +89,6 @@ export class IconTabBarProcessTypeComponent extends IconTabBarBase {
                 isPreviousStepsStrategy
                     ? this.recalculateItemsByPrevArr(extra, amountOfPreviousSteps)
                     : this._recalculateItemsByNextArr(extra, amountOfNextSteps);
-                this.extraBtnDirective?.calculatePosition();
                 this._cd.detectChanges();
             }
         });